# Conflicts:
#	CHANGELOG.md
#	package-lock.json
#	package.json
#	src/app/pages/signing/signing.component.ts
#	yarn.lock

+++ src/app/services/wallet-events.service.ts
@@ -9,17 +9,30 @@ import { AccountService } from '@app/services/account.service';
 import { AccountOverview } from '@app/types/AccountOverview';
 import { SignerService } from '@app/services/signer.service';
 import { AddressBookEntry } from '@app/types/AddressBookEntry';
+import { TldEntry } from '@app/types/TldEntry';
 import { SpyglassService } from '@app/services/spyglass.service';
 import { CurrencyConversionService } from '@app/services/currency-conversion.service';
 import { AuthGuardService } from '../guards/auth-guard';
 import { Router } from '@angular/router';
+import { Datasource, DatasourceService } from '@app/services/datasource.service';
+import { ReceiveSnackbarComponent } from '@app/overlays/snackbar/receive-snackbar.component';
+import { ReceiveService } from '@app/services/receive.service';
 
-const SNACKBAR_DURATION = 3000;
-const SNACKBAR_CLOSE_ACTION_TEXT = 'Dismiss';
+export const SNACKBAR_DURATION = 4000;
+export const SNACKBAR_CLOSE_ACTION_TEXT = 'Dismiss';
 const sortAccounts = (accounts): AccountOverview[] => accounts.sort((a, b) => (a.index < b.index ? -1 : 1));
 
 /** User has request next sequential index be added to the dashboard. */
 export const ADD_NEXT_ACCOUNT_BY_INDEX = new Subject<void>();
+
+/** New Banano Node (URL) has been added to the settings page. */
+export const ADD_RPC_NODE_BY_URL = new Subject<string>();
+
+/** New Spyglass API source (URL) has been added to the settings page. */
+export const ADD_SPYGLASS_API_SOURCE_BY_URL = new Subject<string>();
+
+/** New Banano TLD has been added to the settings page. */
+export const ADD_TLD = new Subject<TldEntry>();
 
 /** New addresses (index) has been added to the dashboard. */
 export const ADD_SPECIFIC_ACCOUNTS_BY_INDEX = new Subject<number[]>();
@@ -29,6 +42,9 @@ export const ATTEMPT_UNLOCK_WALLET_WITH_PASSWORD = new Subject<{ password: strin
 
 /** User wants to unlock the ledger device. */
 export const ATTEMPT_UNLOCK_LEDGER_WALLET = new Subject<void>();
+
+/** All accounts have been loaded. */
+export const AUTO_RECEIVE_ALL = new Subject<void>();
 
 /** Browser supports USB functionality and can be read by the ledger device. */
 export const BROWSER_SUPPORTS_USB = new Subject<void>();
@@ -69,20 +85,14 @@ export const IMPORT_NEW_WALLET_FROM_SECRET = new Subject<{ secret: string; passw
 /** A wallet (previously unlocked with a password) has been logged out. */
 export const LOCK_WALLET = new Subject<void>();
 
-/** An address (index) has been removed from the dashboard. */
-export const REMOVE_ACCOUNTS_BY_INDEX = new Subject<number[]>();
-
 /** User has requested a specific account be refreshed. */
 export const REFRESH_SPECIFIC_ACCOUNT_BY_INDEX = new Subject<number>();
 
 /** User has requested that all loaded indexes be refreshed, checking for receivable transactions and updating account balances. */
 export const REFRESH_DASHBOARD_ACCOUNTS = new Subject<void>();
 
-/** The active wallet has been given an alias. */
-export const RENAME_ACTIVE_WALLET = new Subject<string>();
-
-/** The user wants to manually add an address to their address book. */
-export const UPDATE_ADDRESS_BOOK = new Subject<AddressBookEntry>();
+/** An address (index) has been removed from the dashboard. */
+export const REMOVE_ACCOUNTS_BY_INDEX = new Subject<number[]>();
 
 /** The user wants to manually add an address to their address book. */
 export const REMOVE_ADDRESS_BOOK_ENTRY = new Subject<AddressBookEntry>();
@@ -93,14 +103,38 @@ export const REMOVE_ACTIVE_WALLET = new Subject<void>();
 /** User has opted to delete all locally stored info. */
 export const REMOVE_ALL_WALLET_DATA = new Subject<void>();
 
+/** The active wallet has been given an alias. */
+export const RENAME_ACTIVE_WALLET = new Subject<string>();
+
+/** A Banano Node (URL) has been removed from the settings page. The display order on the settings page matches the order in storage.  */
+export const REMOVE_CUSTOM_RPC_NODE_BY_INDEX = new Subject<number>();
+
+/** A Spyglass API datasource been removed from the settings page. The display order on the settings page matches the order in storage.  */
+export const REMOVE_CUSTOM_SPYGLASS_API_BY_INDEX = new Subject<number>();
+
+export const REMOVE_TLD_BY_NAME = new Subject<string>();
+
 /** User has requested a backup action */
 export const REQUEST_BACKUP_SECRET = new Subject<{ useMnemonic: boolean }>();
 
 /** An account is being added to the dashboard. Can be either true or false. */
 export const SET_DASHBOARD_ACCOUNT_LOADING = new BehaviorSubject<boolean>(true);
 
+/** User has changed which currency they want to use when converting Banano to currency amounts. */
+export const SELECT_LOCALIZATION_CURRENCY = new Subject<string>();
+
+/** Datasource RPC has been updated. */
+export const SELECTED_RPC_DATASOURCE_CHANGE = new Subject<Datasource>();
+
+/** Spyglass API source has been updated. */
+export const SELECTED_SPYGLASS_API_DATASOURCE_CHANGE = new Subject<Datasource>();
+
 /** A transaction has been broadcast onto the network successfully. */
-export const TRANSACTION_COMPLETED_SUCCESS = new Subject<string | undefined>();
+export const TRANSACTION_COMPLETED_SUCCESS = new Subject<{
+    txHash?: string;
+    accountIndex?: number;
+    recipient?: string;
+}>();
 
 /** A wallet (either secret or ledger) has been unlocked. */
 export const UNLOCK_WALLET = new Subject<{ isLedger: boolean; password: string }>();
@@ -108,8 +142,17 @@ export const UNLOCK_WALLET = new Subject<{ isLedger: boolean; password: string }
 /** User has provided an incorrect password to unlock the wallet. */
 export const UNLOCK_WALLET_WITH_PASSWORD_ERROR = new Subject<void>();
 
-/** User has changed which currency they want to use when converting Banano to currency amounts. */
-export const SELECT_LOCALIZATION_CURRENCY = new Subject<string>();
+/** The user wants to manually add an address to their address book. */
+export const UPDATE_ADDRESS_BOOK = new Subject<AddressBookEntry>();
+
+/** User has opted to auto-receive transactions (secret-only) when wallet is unlocked. */
+export const USER_TOGGLE_AUTO_RECEIVE = new Subject<boolean>();
+
+/** User received all incoming transactions via auto-receive overlay. */
+export const USER_COMPLETE_AUTO_RECEIVE = new Subject<void>();
+
+/** User has hit the cancel option from within the auto-receive overlay. */
+export const USER_CANCEL_AUTO_RECEIVE = new Subject<void>();
 
 @Injectable({
     providedIn: 'root',
@@ -117,12 +160,15 @@ export const SELECT_LOCALIZATION_CURRENCY = new Subject<string>();
 export class WalletEventsService {
     store: AppStore;
 
-    private _loadOnlineRepsAndKnownAccounts(): void {
+    private _loadNetworkMetadata(): void {
         void this._accountService.fetchOnlineRepresentatives().then((onlineRepresentatives) => {
             this._appStateService.onlineRepresentatives = onlineRepresentatives;
         });
         void this._accountService.fetchKnownAccounts().then((knownAccounts) => {
             this._appStateService.knownAccounts = knownAccounts;
+        });
+        void this._spyglassService.getRepresentativeScores().then((repScores) => {
+            this._appStateService.repScores = repScores;
         });
     }
 
@@ -133,15 +179,23 @@ export class WalletEventsService {
         private readonly _authGuard: AuthGuardService,
         private readonly _signerService: SignerService,
         private readonly _secretService: SecretService,
+        private readonly _receiveService: ReceiveService,
         private readonly _accountService: AccountService,
         private readonly _spyglassService: SpyglassService,
+        private readonly _datasourceService: DatasourceService,
         private readonly _appStateService: AppStateService,
         private readonly _walletStorageService: WalletStorageService,
         private readonly _currencyConversionService: CurrencyConversionService
-    ) {
+    ) {}
+
+    init(): void {
+        // eslint-disable-next-line no-console
+        console.log('Wallet Events Service Initialized');
+
         // _dispatch initial app state
         this._dispatch({
             activeWallet: undefined,
+            isEnableAutoReceiveFeature: this._walletStorageService.readAutoReceiveFlag(),
             minimumBananoThreshold: this._walletStorageService.readMinimumBananoIncomingThreshold(),
             localCurrencyCode: this._walletStorageService.readLocalizationCurrencyFromLocalStorage(),
             addressBook: this._walletStorageService.readAddressBookFromLocalStorage(),
@@ -149,6 +203,9 @@ export class WalletEventsService {
             localStorageWallets: this._walletStorageService.readWalletsFromLocalStorage(),
             preferredDashboardView: this._walletStorageService.readPreferredDashboardViewFromLocalStorage(),
             idleTimeoutMinutes: this._walletStorageService.readIdleTimeoutMinutes(),
+            customRpcNodeSources: this._walletStorageService.readCustomRpcNodeUrls(),
+            customSpyglassApiSources: this._walletStorageService.readCustomSpyglassUrls(),
+            tlds: this._walletStorageService.readTlds(),
         });
 
         this._appStateService.store.subscribe((store) => {
@@ -158,6 +215,21 @@ export class WalletEventsService {
         ADD_NEXT_ACCOUNT_BY_INDEX.subscribe(() => {
             const nextIndex = this._accountService.findNextUnloadedIndex();
             ADD_SPECIFIC_ACCOUNTS_BY_INDEX.next([nextIndex]);
+        });
+
+        ADD_RPC_NODE_BY_URL.subscribe((url: string) => {
+            this._datasourceService.addCustomDatasource('rpc', url);
+            this._dispatch({ customRpcNodeSources: this.store.customRpcNodeSources.concat(url) });
+        });
+
+        ADD_SPYGLASS_API_SOURCE_BY_URL.subscribe((url: string) => {
+            this._datasourceService.addCustomDatasource('spyglass', url);
+            this._dispatch({ customSpyglassApiSources: this.store.customSpyglassApiSources.concat(url) });
+        });
+
+        ADD_TLD.subscribe((tld: TldEntry) => {
+            this.store.tlds[tld.name] = tld.account;
+            this._dispatch({ tlds: this.store.tlds });
         });
 
         ADD_SPECIFIC_ACCOUNTS_BY_INDEX.subscribe(async (indexes: number[]) => {
@@ -178,6 +250,7 @@ export class WalletEventsService {
                 isLoadingAccounts: false,
                 accounts: sortAccounts(accounts),
             });
+            AUTO_RECEIVE_ALL.next();
         });
 
         ATTEMPT_UNLOCK_LEDGER_WALLET.subscribe(async () => {
@@ -201,6 +274,26 @@ export class WalletEventsService {
                 console.error(err);
                 UNLOCK_WALLET_WITH_PASSWORD_ERROR.next();
             }
+        });
+
+        AUTO_RECEIVE_ALL.subscribe(() => {
+            if (!this.store.isEnableAutoReceiveFeature) {
+                return;
+            }
+            if (!this.store.hasUnlockedSecret) {
+                return;
+            }
+            if (this.store.isAutoReceivingTransactions) {
+                return;
+            }
+            const blocks = this._appStateService.getAllReceivableBlocks();
+            if (blocks.length === 0) {
+                return;
+            }
+            this._snackbar.openFromComponent(ReceiveSnackbarComponent);
+            this._dispatch({
+                isAutoReceivingTransactions: true,
+            });
         });
 
         BROWSER_SUPPORTS_USB.subscribe(() => {
@@ -296,19 +389,23 @@ export class WalletEventsService {
             if (indexes.length === 0) {
                 indexes.push(0);
             }
-            this._loadOnlineRepsAndKnownAccounts();
+            this._loadNetworkMetadata();
             ADD_SPECIFIC_ACCOUNTS_BY_INDEX.next(indexes);
             SELECT_LOCALIZATION_CURRENCY.next(this.store.localCurrencyCode);
         });
 
         REFRESH_SPECIFIC_ACCOUNT_BY_INDEX.subscribe(async (index) => {
             const newAccount = await this._accountService.fetchAccount(index);
+            if (!newAccount) {
+                return;
+            }
             const accounts = this._accountService.removeAccounts([index]);
             accounts.push(newAccount);
             this._dispatch({
                 accounts: sortAccounts(accounts),
                 totalBalance: this._accountService.calculateLoadedAccountsTotalBalance(accounts),
             });
+            AUTO_RECEIVE_ALL.next();
         });
 
         REMOVE_ACCOUNTS_BY_INDEX.subscribe((indexes: number[]) => {
@@ -326,16 +423,6 @@ export class WalletEventsService {
         REMOVE_ADDRESS_BOOK_ENTRY.subscribe((entry: AddressBookEntry) => {
             const addressBook = this.store.addressBook;
             addressBook.delete(entry.account);
-            this._dispatch({ addressBook });
-        });
-
-        UPDATE_ADDRESS_BOOK.subscribe((entry: AddressBookEntry) => {
-            const addressBook = this.store.addressBook;
-            if (entry.account === entry.name) {
-                addressBook.delete(entry.account);
-            } else {
-                addressBook.set(entry.account, entry.name);
-            }
             this._dispatch({ addressBook });
         });
 
@@ -367,6 +454,23 @@ export class WalletEventsService {
             LOCK_WALLET.next();
         });
 
+        REMOVE_CUSTOM_RPC_NODE_BY_INDEX.subscribe((index) => {
+            this._dispatch({
+                customRpcNodeSources: this._datasourceService.removeCustomRpcSource(index),
+            });
+        });
+
+        REMOVE_CUSTOM_SPYGLASS_API_BY_INDEX.subscribe((index) => {
+            this._dispatch({
+                customSpyglassApiSources: this._datasourceService.removeCustomSpyglassSource(index),
+            });
+        });
+
+        REMOVE_TLD_BY_NAME.subscribe((name) => {
+            delete this.store.tlds[name];
+            this._dispatch({ tlds: this.store.tlds });
+        });
+
         REQUEST_BACKUP_SECRET.subscribe(async (data) => {
             if (data.useMnemonic) {
                 const mnemonic = await this._secretService.getActiveWalletMnemonic();
@@ -377,18 +481,26 @@ export class WalletEventsService {
             }
         });
 
+        SELECT_LOCALIZATION_CURRENCY.subscribe(async (localCurrencyCode: string) => {
+            const priceDataUSD = await this._spyglassService.getBananoPriceRelativeToBitcoin();
+            const localCurrencyConversionRate = this._currencyConversionService.convertToUSD(localCurrencyCode);
+            this._dispatch({ localCurrencyCode, localCurrencyConversionRate, priceDataUSD });
+        });
+
         SET_DASHBOARD_ACCOUNT_LOADING.subscribe((isLoadingAccounts) => {
             this._dispatch({ isLoadingAccounts });
         });
 
-        SELECT_LOCALIZATION_CURRENCY.subscribe(async (localCurrencyCode: string) => {
-            const priceDataUSD = await this._spyglassService.getBananoPriceRelativeToBitcoin();
-            const localCurrencyConversionRate = await this._currencyConversionService.convertToUSD(localCurrencyCode);
-            this._dispatch({ localCurrencyCode, localCurrencyConversionRate, priceDataUSD });
-        });
-
-        TRANSACTION_COMPLETED_SUCCESS.subscribe(() => {
-            REFRESH_DASHBOARD_ACCOUNTS.next();
+        TRANSACTION_COMPLETED_SUCCESS.subscribe((data) => {
+            if (this.store.hasUnlockedSecret && data.accountIndex !== undefined) {
+                // Do not refresh for ledger devices; the signer is already in progress and subsequent calls will fail at this point.
+                REFRESH_SPECIFIC_ACCOUNT_BY_INDEX.next(data.accountIndex);
+                this.store.accounts.map((account) => {
+                    if (account.fullAddress === data.recipient) {
+                        REFRESH_SPECIFIC_ACCOUNT_BY_INDEX.next(account.index);
+                    }
+                });
+            }
         });
 
         UNLOCK_WALLET.subscribe((data) => {
@@ -404,6 +516,44 @@ export class WalletEventsService {
             });
             REFRESH_DASHBOARD_ACCOUNTS.next();
         });
+
+        UPDATE_ADDRESS_BOOK.subscribe((entry: AddressBookEntry) => {
+            const addressBook = this.store.addressBook;
+            if (entry.account === entry.name) {
+                addressBook.delete(entry.account);
+            } else {
+                addressBook.set(entry.account, entry.name);
+            }
+            this._dispatch({ addressBook });
+        });
+
+        USER_TOGGLE_AUTO_RECEIVE.subscribe((isEnabled: boolean) => {
+            this._dispatch({
+                isEnableAutoReceiveFeature: isEnabled,
+            });
+            if (isEnabled) {
+                AUTO_RECEIVE_ALL.next();
+            }
+        });
+
+        USER_COMPLETE_AUTO_RECEIVE.subscribe(() => {
+            this._dispatch({
+                isAutoReceivingTransactions: false,
+            });
+            setTimeout(() => {
+                this._snackbar.dismiss();
+            }, SNACKBAR_DURATION);
+            REFRESH_DASHBOARD_ACCOUNTS.next();
+            TRANSACTION_COMPLETED_SUCCESS.next({});
+        });
+
+        USER_CANCEL_AUTO_RECEIVE.subscribe(() => {
+            this._receiveService.stopReceive();
+            this._snackbar.dismiss();
+            this._dispatch({
+                isAutoReceivingTransactions: false,
+            });
+        });
     }
 
     /** Broadcasts an updated app state. */
@@ -417,9 +567,14 @@ export class WalletEventsService {
             newData.addressBook ||
             newData.localCurrencyCode ||
             newData.preferredDashboardView ||
+            newData.customRpcNodeSources ||
+            newData.customSpyglassApiSources ||
+            newData.tlds ||
+            newData.isEnableAutoReceiveFeature !== undefined || // Boolean
             newData.minimumBananoThreshold !== undefined // Can be 0.
         ) {
             this._appStateService.appLocalStorage.next({
+                customRpcNodeSources: newData.customRpcNodeSources,
                 minimumBananoThreshold: newData.minimumBananoThreshold,
                 preferredDashboardView: newData.preferredDashboardView,
                 localizationCurrencyCode: newData.localCurrencyCode,
@@ -427,6 +582,9 @@ export class WalletEventsService {
                 activeWallet: newData.activeWallet,
                 localStorageWallets: newData.localStorageWallets,
                 idleTimeoutMinutes: newData.idleTimeoutMinutes,
+                isEnableAutoReceiveFeature: newData.isEnableAutoReceiveFeature,
+                customSpyglassApiSources: newData.customSpyglassApiSources,
+                tlds: newData.tlds,
             });
         }
     }
